% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLE2-package.R
\docType{package}
\name{HLE2-package}
\alias{HLE2}
\alias{HLE2-package}
\title{HLE2: Hodges-Lehmann Estimate of Shift}
\description{
Calculates the Hodges-Lehmann estimator of shift using two different internal methods. First, the "outer" R function and secondly using C++ for performance
}
\keyword{internal}
